Add auto-refresh for "last update ago" date in value card widgets
  Apply periodic date refresh logic to tb-value-card-widget and
  tb-aggregated-value-card-widget. Both widgets display a "Last update
  X ago" timestamp that previously only updated when new telemetry
  arrived. The interval subscription now re-evaluates the relative time
  on every tick (interval driven by dynamicPageLinkRefreshIntervalSec
  from system params, defaulting to 60s), keeping the label accurate
  without waiting for a data push.
