+++ ui-ngx/src/app/modules/home/components/widget/lib/cards/aggregated-value-card-widget.component.ts
@@ -18,6 +18,7 @@ import {
   AfterViewInit,
   ChangeDetectorRef,
   Component,
+  DestroyRef,
   ElementRef,
   Input,
   OnDestroy,
@@ -26,6 +27,7 @@ import {
   TemplateRef,
   ViewChild
 } from '@angular/core';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import {
   aggregatedValueCardDefaultSettings,
   AggregatedValueCardKeyPosition,
@@ -35,7 +37,7 @@ import {
   getTsValueByLatestDataKey
 } from '@home/components/widget/lib/cards/aggregated-value-card.models';
 import { WidgetContext } from '@home/models/widget-component.models';
-import { Observable } from 'rxjs';
+import { interval, Observable } from 'rxjs';
 import {
   autoDateFormat,
   backgroundStyle,
@@ -49,6 +51,7 @@ import {
 import { DataKey } from '@shared/models/widget.models';
 import { formatNumberValue, formatValue, isDefined, isDefinedAndNotNull, isNumeric } from '@core/utils';
 import { map } from 'rxjs/operators';
+import { getCurrentAuthState } from '@core/auth/auth.selectors';
 import { ImagePipe } from '@shared/pipe/image.pipe';
 import { DomSanitizer } from '@angular/platform-browser';
 import { TbTimeSeriesChart } from '@home/components/widget/lib/chart/time-series-chart';
@@ -121,7 +124,8 @@ export class AggregatedValueCardWidgetComponent implements OnInit, AfterViewInit
   constructor(private imagePipe: ImagePipe,
               private sanitizer: DomSanitizer,
               private renderer: Renderer2,
-              private cd: ChangeDetectorRef) {
+              private cd: ChangeDetectorRef,
+              private destroyRef: DestroyRef) {
   }
 
   ngOnInit(): void {
@@ -167,6 +171,18 @@ export class AggregatedValueCardWidgetComponent implements OnInit, AfterViewInit
     this.dateFormat = DateFormatProcessor.fromSettings(this.ctx.$injector, this.settings.dateFormat);
     this.dateStyle = textStyle(this.settings.dateFont);
     this.dateColor = this.settings.dateColor;
+
+    if (this.showDate && this.settings.dateFormat?.lastUpdateAgo) {
+      const refreshIntervalSec = getCurrentAuthState(this.ctx.store)?.dynamicPageLinkRefreshIntervalSec || 60;
+      interval(refreshIntervalSec * 1000).pipe(
+        takeUntilDestroyed(this.destroyRef)
+      ).subscribe(() => {
+        if (this.lastUpdateTs) {
+          this.dateFormat.update(this.lastUpdateTs);
+          this.cd.detectChanges();
+        }
+      });
+    }
 
     this.backgroundStyle$ = backgroundStyle(this.settings.background, this.imagePipe, this.sanitizer);
     this.overlayStyle = overlayStyle(this.settings.background.overlay);
