+++ ui-ngx/src/app/modules/home/components/widget/lib/cards/value-chart-card-widget.component.scss
@@ -50,9 +50,6 @@
     .tb-value-chart-card-value {
       white-space: nowrap;
     }
-    .tb-value-chart-card-date {
-      white-space: nowrap;
-    }
     .tb-value-chart-card-chart {
       flex: 1;
       min-width: 0;
