+++ ui-ngx/src/app/modules/home/components/widget/lib/cards/value-chart-card-widget.models.ts
@@ -19,9 +19,7 @@ import {
   BackgroundType,
   ColorSettings,
   constantColor,
-  DateFormatSettings,
-  Font,
-  lastUpdateAgoDateFormat
+  Font
 } from '@shared/models/widget-settings.models';
 
 export enum ValueChartCardLayout {
@@ -51,10 +49,6 @@ export interface ValueChartCardWidgetSettings {
   showValue: boolean;
   valueFont: Font;
   valueColor: ColorSettings;
-  showDate: boolean;
-  dateFormat: DateFormatSettings;
-  dateFont: Font;
-  dateColor: ColorSettings;
   background: BackgroundSettings;
   padding: string;
 }
@@ -72,17 +66,6 @@ export const valueChartCardDefaultSettings: ValueChartCardWidgetSettings = {
     lineHeight: '32px'
   },
   valueColor: constantColor('rgba(0, 0, 0, 0.87)'),
-  showDate: true,
-  dateFormat: lastUpdateAgoDateFormat(),
-  dateFont: {
-    family: 'Roboto',
-    size: 12,
-    sizeUnit: 'px',
-    style: 'normal',
-    weight: '400',
-    lineHeight: '16px'
-  },
-  dateColor: constantColor('rgba(0, 0, 0, 0.38)'),
   background: {
     type: BackgroundType.color,
     color: '#fff',
