+++ ui-ngx/src/app/modules/home/components/widget/lib/settings/cards/value-chart-card-widget-settings.component.html
@@ -50,19 +50,6 @@
         </tb-color-settings>
       </div>
     </div>
-    <div class="tb-form-row column-xs">
-      <mat-slide-toggle class="mat-slide" formControlName="showDate">
-        {{ 'widgets.value-chart-card.date' | translate }}
-      </mat-slide-toggle>
-      <div class="flex flex-row items-center justify-start gap-2">
-        <tb-date-format-select class="flex-1" formControlName="dateFormat"></tb-date-format-select>
-        <tb-font-settings formControlName="dateFont"
-                          [previewText]="datePreviewFn">
-        </tb-font-settings>
-        <tb-color-settings formControlName="dateColor" settingsKey="{{'widgets.value-chart-card.date' | translate }}">
-        </tb-color-settings>
-      </div>
-    </div>
     <div class="tb-form-row space-between">
       <div>{{ 'widgets.background.background' | translate }}</div>
       <tb-background-settings formControlName="background">
